requirements.txt
flask==2.2.5
pymongo==4.3.3
redis==4.5.4
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.3
PyPDF2==3.0.1
docx2txt==0.8
sentence-transformers==4.1.0 